Actualizacion del programa - Casi terminado

+++ src/main/java/t4p/persistence/TweetDao.java
@@ -9,7 +9,7 @@ public interface TweetDao {
 
 	public abstract Collection<Tweet> findAll();
 
-	public abstract Tweet findById(Long id);
+	public abstract Tweet findById(int id);
 
 	public abstract boolean delete(Tweet tweet);
